requirements.txt
# User code dependencies
pandas
numpy
gcsfs

# Parse user config
pyyaml

# KFP Building
kfp
docker
jinja2

# Running an example pipeline
google-cloud-aiplatform